requirements.txt
paho-mqtt>=1.6.0
Pillow>=9.0.0
pynput>=1.7.0

